changed notes and attachments permissions in a way that only the uploader can edit or delete them

+++ resources/views/notes/show.blade.php
@@ -20,6 +20,7 @@
           <h3>{{ $note->text }}</h3>
         </ul>
 
+            @if($user == $note->user)
               <!-- Right Side Of Navbar -->
               <ul class="nav navbar-nav navbar-left" style="margin-top: 20px;margin-left: 20px;">
                   {!! Form::open(array('class' => 'form-inline', 'method' => 'DELETE', 'onsubmit' => 'return ConfirmDelete()' ,'route' => array('fleet_lists.aircrafts.notes.destroy', $fleet_list->id, $aircraft->id, $note->id))) !!}
@@ -29,6 +30,7 @@
                   {!! Form::button('<span class="glyphicon glyphicon-trash"></span> Delete', array('class' => 'btn btn-danger', 'type' => 'submit')) !!}
                   {!! Form::close() !!}
               </ul>
+            @endif
           </div>
       </div>
   </nav>
